Conditionally compile pipeline debug infrastructure.
Use go test -tags=debug ./pipeline/ ........ to enable them.
